Updated RemoveTaskAction to either update lastCOmpletedDate or remove all tasks

+++ src/main/java/com/tasktracker/controller/RemoveTaskAction.java
@@ -9,7 +9,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
+import java.time.LocalDate;
 
 @WebServlet(
         urlPatterns = {"/users/removeTaskAction"}
@@ -19,17 +21,30 @@ public class RemoveTaskAction extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        HttpSession session = req.getSession();
+
         //Get task to be removed
         String taskId = req.getParameter("id");
         GenericDao taskDao = new GenericDao(Task.class);
         Task taskToRemove = (Task) taskDao.getById(Integer.parseInt(taskId));
 
-        //
+        //If it's a 'once' task, delete it
+        if (req.getParameter("frequency").equals("once")) {
+            taskDao.delete(taskToRemove);
+        }
 
-        //Delete task
-        taskDao.delete(taskToRemove);
+        //If it's a recurring task, determine which instances the user wants to remove
+        if (!req.getParameter("frequency").equals("once")) {
+            if (req.getParameter("instances").equals("onlyThis")) {
+                taskToRemove.setLastDateCompleted((LocalDate)session.getAttribute("now"));
+                taskDao.saveOrUpdate(taskToRemove);
 
-        //TODO: Message that task was successfully deleted?
+            } else if (req.getParameter("instances").equals("allInstances")) {
+                taskDao.delete(taskToRemove);
+            }
+        }
+
+        //TODO: Message that task was successfully updated?
 
         resp.sendRedirect(req.getContextPath() + "/users/viewPlanner");
     }
